[Mod]修改WebTrader样例支持无人值守模式
